Add program base offset option
The program base offset can now be set.

For example, previously if there was an ORG 0100H directive at the start of the program, the emitted hex file would contain 256 0xFF bytes at the start.
Now, if the option --base=0x100 is used, the hex file will begin with code assembled at address 0x100.

In this scenario, attempting to assemble any code below the address of 0x100 will result in an error.